pyproject.toml
[project]
name = "memoir"
version = "0.1.0"
description = "AI Long-term Memory Framework - File-based memory with transparency and associative retrieval"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "chromadb>=0.4.22",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "pypdf>=3.17.4",
    "python-docx>=1.1.0",
    "pillow>=10.2.0",
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
